flask/requirements.txt
flask
flask-restful
python-dotenv
psycopg2-binary

